refactor [KVCacheManager]: modularize locking and page allocation logic by introducing dedicated classes for locks and page management
